chore: bgp integration tests
