+++ .github/workflows/main.yml
@@ -68,6 +68,9 @@ jobs:
   integration-tests-bgp-5g:
     needs: [image-build, image-build-tester]
     uses: ./.github/workflows/integration-tests-bgp-5g.yaml
+  integration-tests-bgp-4g:
+    needs: [image-build, image-build-tester]
+    uses: ./.github/workflows/integration-tests-bgp-4g.yaml
   integration-tests-rolling-upgrade:
     needs: [image-build]
     uses: ./.github/workflows/integration-tests-rolling-upgrade.yaml
