+++ integration/core_tester_test.go
@@ -30,6 +30,7 @@ var scenariosSkipped = map[string]string{
 	"gnb/connectivity_expect_blocked_ipv6": "test-only harness; requires a pre-installed deny rule",
 	"gnb/connectivity_expect_allowed_ipv6": "test-only harness; minimal allow-path",
 	"gnb/session_hold":                     "long-lived session for BGP tests; covered by TestIntegration5GBGP",
+	"s1enb/session_hold":                   "long-lived session for BGP tests; covered by TestIntegration4GBGP",
 	"gnb/nat_checksum":                     "capture-driven harness; covered by TestIntegration5GUPFNATChecksum",
 	"s1enb/connectivity_expect_allowed":    "test-only harness; driven by TestIntegration4GNetworkRules",
 	"s1enb/connectivity_expect_blocked":    "test-only harness; driven by TestIntegration4GNetworkRules",
